+++ sentry/src/main/java/io/sentry/protocol/MeasurementValue.java
@@ -1,0 +1,13 @@
+package io.sentry.protocol;
+
+import org.jetbrains.annotations.ApiStatus;
+
+@ApiStatus.Internal
+public final class MeasurementValue {
+  @SuppressWarnings("UnusedVariable")
+  private final float value;
+
+  public MeasurementValue(final float value) {
+    this.value = value;
+  }
+}