+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Feat: Measure app start time (#1487)
 * Feat: Automatic breadcrumbs logging for fragment lifecycle (#1522) 
 
 ## 5.0.1